refactor(auv-tracing-driver): move durable recording modules
