+++ crates/auv-tracing-driver/Cargo.toml
@@ -14,6 +14,9 @@ repository.workspace = true
 [dependencies]
 serde.workspace = true
 serde_json.workspace = true
-tokio = { version = "1", features = ["sync"] }
+tokio = { version = "1", features = ["macros", "net", "rt-multi-thread", "sync"] }
 reqwest.workspace = true
 tracing = "0.1"
+
+[dev-dependencies]
+axum = "0.8"
