+++ crates/auv-tracing-driver/src/recording/update.rs
